Transfer spy stub docs and hints into implementations
