+++ crudclient/testing/spy/base.py
@@ -1,3 +1,9 @@
+"""Base components for the Test Spy pattern.
+
+This module defines ``SpyBase`` which provides the core functionality for
+recording method calls in test doubles.
+"""
+
 import warnings
 from typing import Any, Dict, List, Optional, Tuple
 
@@ -6,7 +12,15 @@ from .method_call import MethodCall
 
 
 class SpyBase:
+    """Base implementation for the **Test Spy pattern**.
+
+    A Test Spy records information about how it was called during test
+    execution.  Concrete spy classes should inherit from this base to leverage
+    the call recording and verification infrastructure provided here.
+    """
+
     def __init__(self) -> None:
+        """Initialize the spy base."""
         self.calls: List[MethodCall] = []
 
     def _record_call(
@@ -17,15 +31,18 @@ class SpyBase:
         return_value: Optional[Any] = None,
         exception: Optional[Exception] = None,
     ) -> None:
+        """Record a method call."""
         call = MethodCall(method_name, args, kwargs, return_value, exception)
         self.calls.append(call)
 
     def _format_args_string(self, *args: Any, **kwargs: Any) -> str:
+        """Format arguments as a string."""
         args_str = ", ".join(str(arg) for arg in args)
         kwargs_str = ", ".join(f"{key}={value}" for key, value in kwargs.items())
         return ", ".join(filter(None, [args_str, kwargs_str]))
 
     def verify_called(self, method_name: str) -> None:
+        """Verify that a method was called."""
         for call in self.calls:
             if call.method_name == method_name:
                 return
@@ -33,11 +50,13 @@ class SpyBase:
         raise SpyError(f"Method {method_name} was not called")
 
     def verify_not_called(self, method_name: str) -> None:
+        """Verify that a method was not called."""
         for call in self.calls:
             if call.method_name == method_name:
                 raise SpyError(f"Method {method_name} was called")
 
     def verify_called_with(self, method_name: str, *args: Any, **kwargs: Any) -> None:
+        """Verify that a method was called with specific arguments."""
         for call in self.calls:
             if call.method_name == method_name:
                 # Check positional arguments
@@ -54,19 +73,23 @@ class SpyBase:
         raise SpyError(f"Method {method_name} was not called with arguments ({all_args})")
 
     def verify_call_count(self, method_name: str, count: int) -> None:
+        """Verify that a method was called a specific number of times."""
         actual_count = sum(1 for call in self.calls if call.method_name == method_name)
 
         if actual_count != count:
             raise SpyError(f"Method {method_name} was called {actual_count} times, expected {count} times")
 
     def get_calls(self, method_name: str) -> List[MethodCall]:
+        """Get all calls to a specific method."""
         return [call for call in self.calls if call.method_name == method_name]
 
     def clear_calls(self) -> None:
+        """Clear all recorded calls."""
         self.calls = []
 
     # Deprecated methods for backward compatibility
     def assert_called(self, method_name: str) -> None:
+        """Assert that a method was called (deprecated)."""
         warnings.warn(
             "assert_called is deprecated and will be removed in a future version. Use verify_called instead.",
             DeprecationWarning,
@@ -78,6 +101,7 @@ class SpyBase:
             raise AssertionError(str(e))
 
     def assert_not_called(self, method_name: str) -> None:
+        """Assert that a method was not called (deprecated)."""
         warnings.warn(
             "assert_not_called is deprecated and will be removed in a future version. Use verify_not_called instead.",
             DeprecationWarning,
@@ -89,6 +113,7 @@ class SpyBase:
             raise AssertionError(str(e))
 
     def assert_called_with(self, method_name: str, *args: Any, **kwargs: Any) -> None:
+        """Assert that a method was called with specific arguments (deprecated)."""
         warnings.warn(
             "assert_called_with is deprecated and will be removed in a future version. Use verify_called_with instead.",
             DeprecationWarning,
@@ -100,6 +125,7 @@ class SpyBase:
             raise AssertionError(str(e))
 
     def assert_call_count(self, method_name: str, count: int) -> None:
+        """Assert that a method was called a specific number of times (deprecated)."""
         warnings.warn(
             "assert_call_count is deprecated and will be removed in a future version. Use verify_call_count instead.",
             DeprecationWarning,
