+++ crudclient/testing/spy/method_call.py
@@ -1,11 +1,15 @@
+"""Data structure for representing a recorded method call."""
+
 from typing import Any, Dict, Optional, Tuple
 
 
 class MethodCall:
+    """Record of a method call for verification."""
 
     def __init__(
         self, method_name: str, args: Tuple[Any, ...], kwargs: Dict[str, Any], return_value: Any = None, exception: Optional[Exception] = None
     ):
+        """Initialize a method call record."""
         self.method_name = method_name
         self.args = args
         self.kwargs = kwargs
@@ -13,6 +17,7 @@ class MethodCall:
         self.exception = exception
 
     def __repr__(self) -> str:
+        """String representation of the method call."""
         args_str = ", ".join([repr(arg) for arg in self.args])
         kwargs_str = ", ".join([f"{k}={repr(v)}" for k, v in self.kwargs.items()])
         all_args = ", ".join(filter(None, [args_str, kwargs_str]))
